Extract PstFlowComputer
